refactor: move `EVMAggregateExec` to its own file
